Add producer to pipeline
Added producer which will write to the first topic in the arguments

Tweaked time calculations in publish_aggregations to get previous window
(was subtracting a negative)

Tweaked metric spec to allow float64 timestamps (found in devstack
from python api)

+++ models/metric.go
@@ -17,7 +17,7 @@ package models
 type Metric struct {
 	Name       string            `json:"name"`
 	Dimensions map[string]string `json:"dimensions"`
-	Timestamp  int64             `json:"timestamp"`
+	Timestamp  float64             `json:"timestamp"`
 	Value      float64           `json:"value"`
 	ValueMeta  map[string]string `json:"value_meta"`
 }
